requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==1.4.18
urllib3==2.0.2
Werkzeug==2.0.1
